Configure OpenFeign - created new module clients that store fraud interface. Record FraudCheckResponse has been moved to the clients module in the OrderApplication has been enabled OpenFeign clients.
