+++ clients/src/main/java/com/bartoszmarkiewicz/clients/fraud/FraudCheckResponse.java
@@ -1,4 +1,4 @@
-package com.bartoszmarkiewicz.order;
+package com.bartoszmarkiewicz.clients.fraud;
 
 public record FraudCheckResponse(boolean isFraudster) {
