+++ fraud/src/main/java/com/bartoszmarkiewicz/fraud/FraudApplication.java
@@ -4,6 +4,7 @@ package com.bartoszmarkiewicz.fraud;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
 @EnableEurekaClient
