+++ order/src/main/java/com/bartoszmarkiewicz/order/OrderApplication.java
@@ -4,9 +4,13 @@ package com.bartoszmarkiewicz.order;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
 @EnableEurekaClient
+@EnableFeignClients(
+        basePackages = "com.bartoszmarkiewicz.clients"
+)
 public class OrderApplication {
     public static void main(String[] args) {
         SpringApplication.run(OrderApplication.class, args);
