+++ pom.xml
@@ -15,6 +15,7 @@
     <module>inventory</module>
     <module>eureka-server</module>
     <module>fraud</module>
+      <module>clients</module>
   </modules>
 
   <properties>
@@ -53,6 +54,10 @@
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-test</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.cloud</groupId>
+      <artifactId>spring-cloud-starter-openfeign</artifactId>
     </dependency>
   </dependencies>
